pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repo_explainer_agent"
version = "0.1.0"
description = "A simple AI agent analyzing the contents of a (Git) repository. "
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name="Andrei Matveyeu", email="andrei@ideabulbs.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache License 2.0",
    "Operating System :: OS Independent",
]
dependencies = [
    "toml",
    "langgraph==0.4.5",
    "langchain==0.3.25",
    "langchain-community==0.3.24",
    "google-generativeai",
    "GitPython",
]

[tool.setuptools]
package-dir = {"repo_explainer_agent" = "src"}
# This maps the package named 'repo_explainer_agent' to the 'src' directory.
# Setuptools will then look for an __init__.py in 'src' (which we created)
# and include modules like 'graph.py' and subpackages like 'core', 'nodes', 'utils'
# as part of the 'repo_explainer_agent' package.
packages = ["repo_explainer_agent"]


[project.urls]
"Homepage" = "https://github.com/andreimatveyeu/repo_explainer_agent"
"Bug Tracker" = "https://github.com/andreimatveyeu/repo_explainer_agent/issues"
